Update README.md
I did a previous commit, but it was screwed up, so I did it again.  The team projects are found in the v2_id field as opposed to the id field of the attributes.  So if you want to be able to access team projects, the v2_id is the object to reference in the value_template.